# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: API de aprendizaje OpenAPI
  version: 1.0.0
  description: API de ejemplo para aprender OpenAPI con Node.js

paths:
  /:
    get:
      summary: Comprueba que la API está activa
      responses:
        '200':
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: API funcionando correctamente

  /movies:
    get:
      summary: Obtiene la lista de películas
      description: Devuelve un array con las películas disponibles
      responses:
        '200':
          description: Lista de películas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    title:
                      type: string
                      example: Inception
                    director:
                      type: string
                      example: Christopher Nolan

    post:
      summary: Crea una nueva película
      description: Añade una nueva película a la colección
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - director
              properties:
                title:
                  type: string
                  example: The Prestige
                director:
                  type: string
                  example: Christopher Nolan
      responses:
        '201':
          description: Película creada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 3
                  title:
                    type: string
                    example: The Prestige
                  director:
                    type: string
                    example: Christopher Nolan
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Los campos title y director son obligatorios

  /movies/{id}:
    get:
      summary: Obtiene una película por su identificador
      description: Devuelve una película concreta a partir de su id
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la película
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Película encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  title:
                    type: string
                    example: Inception
                  director:
                    type: string
                    example: Christopher Nolan
        '404':
          description: Película no encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Película no encontrada